+++ packages/appimage/CHANGELOG.md
@@ -1,5 +1,13 @@
 # appimage
 
+## 1.0.2
+
+### Patch Changes
+
+- [#109](https://github.com/electron-userland/electron-builder-binaries/pull/109) [`6858c94`](https://github.com/electron-userland/electron-builder-binaries/commit/6858c945046b5034395bdac48cbdf9458ba70aa0) Thanks [@mmaietta](https://github.com/mmaietta)! - fix: compile mksquashfs and desktop-file-validate for mac from source and copy needed dylibs
+
+- [#117](https://github.com/electron-userland/electron-builder-binaries/pull/117) [`6cde90e`](https://github.com/electron-userland/electron-builder-binaries/commit/6cde90e200abbb42016af02c2322650994d25465) Thanks [@mmaietta](https://github.com/mmaietta)! - fix(appimage): auto remove quarantine bit on mac when executing via wrapper script
+
 ## 1.0.1
 
 ### Patch Changes